// generated/src/app/api-models/aucls-setup.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AuclsSetup {
  seqAuclsJobId: string;
  description: string;
  comments: string;
  jobRunOptionCode: number;
  jobActionCode: number;
  jobStatusCode: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}